requirements.txt
grpcio
grpcio-tools
googleapis-common-protos
requests
sox
tqdm
